Grade with float numbers

+++ src/parsers/grade-grid.parser.ts
@@ -25,7 +25,7 @@ function disciplineGradeRows(rows: HTMLElement[]): DisciplineGrade {
   let [partialGradesHeader, partialGradesContent] = partialGradesCell.querySelectorAll('tr');
   let partialGradeNames = partialGradesHeader.querySelectorAll('th').map(x => x.text.trim());
   let partialGradeContents = partialGradesContent.querySelectorAll('td').map(x => {
-    let n = parseInt(x.text);
+    let n = parseFloat(x.text);
     if (isNaN(n)) return null;
     return n;
   });
